requirements.txt
PyQt5>=5.15
openai>=1.0
mss>=9.0
Pillow>=10.0
numpy>=1.24
keyboard>=0.13
